change cost coverage test to monitor zero-signed fee orders
